requirements.txt
# Transit Sentinel - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pydantic>=2.5.0
pyyaml>=6.0
requests>=2.31.0

# GTFS-Realtime Protocol Buffers
gtfs-realtime-bindings>=1.0.0
protobuf>=4.25.0

# Streaming (Kafka/Redpanda)
kafka-python>=2.0.2

# Analytics
duckdb>=0.10.0
pyarrow>=15.0.0

# Orchestration
dagster>=1.6.0
dagster-webserver>=1.6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Development & Testing
pytest>=7.4.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0